instagram.com, aarif_mindblowing, siacheap007@gmail.com
these are the data i put and " iasnaisratica!1A " this is the password it has generated.

do you thing my user will be able to remember this password? do it look like that the password is generated for those data inputs??.

just use keys directly from my input. and create strong password